move shared logic and const to utils and consts
